feat(agent): add a copy button to every chat message
Add a copy-to-clipboard control to user messages and reuse it for
assistant messages. The shared MessageCopyAction follows the app-wide
copy affordance — the duplicate glyph briefly swaps to a success
checkmark after copying — replacing the assistant toolbar's plain copy.
